Run multiple benchmarks in sequence
Created a playbook that can run multiple benchmarks in sequence. The
playbook assumes the instances are available to run.

We start/stop the Hyperfoil controller, and start/stop the server for
each benchmark executed. The playbook requires an input file detailing
all the tests to run, with details about the Hyperfoil benchmark file
and which cache configuration to utilize.